+++ packages/rust/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased (rust-lib@v1.0.3)
+
+ - added a new function `get_schwa_status_for_script` to get the schwa deletion characerstic of the script provided.
+
 ## rust-lib@v1.0.2
 
  - *New Mappings* for Romanized